requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect==1.3.0
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
lxml==5.3.1
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.4
openai==1.66.3
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
safetensors==0.5.3
setuptools==76.1.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
sympy==1.13.1
textblob==0.19.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
